move linking to a node lib adapter option
